Make a module for frame options
