+++ src/util-frame-options.ts
@@ -1,0 +1,11 @@
+import { FrameOptions, FRAME_OPTIONS } from './definitions/FrameOptions'
+import { isKeyOf } from './util'
+
+export function getFrameOptions(frameIdentifier: string): FrameOptions {
+    if (isKeyOf(frameIdentifier, FRAME_OPTIONS)) {
+        return FRAME_OPTIONS[frameIdentifier]
+    }
+    return {
+        multiple: false
+    }
+}
